Add all instancetypes embed
